Add off-by-one module count tests.
