+++ src/cennznut/index.spec.js
@@ -45,7 +45,7 @@ describe("Encode CENNZnut", () => {
 
   it("should decode Rust generated permissions", () => {
     const permissions = new Uint8Array([
-      0, 0, 64, 160, 109, 111, 100, 117, 108, 101, 95, 116, 101, 115, 116, 0, 0, 0, 0, 0, 0,
+      0, 0, 128, 160, 109, 111, 100, 117, 108, 101, 95, 116, 101, 115, 116, 0, 0, 0, 0, 0, 0,
       0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 138, 128, 0, 128, 109, 101, 116, 104,
       111, 100, 95, 116, 101, 115, 116, 50, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
       0, 0, 0, 0, 130, 128, 0, 0, 128, 109, 101, 116, 104, 111, 100, 95, 116, 101, 115, 116,
